Added Drive Controls+Logic
In Drive, I added all the logic to use Tank Drive, and an init method to make sure the motors are stopped while the robot is turning on.
In Actuators, I added the drive motors, and set 1 more on each side to follow the other, such that it will always have the same amount of power going to them both.
In Constants, I added various Contants that were necessary.
In Robot, I added some initilzation logic, and also the drive controls.